Separate `timelines.base.component` into modules
