+++ tilia/timelines/base/component/__init__.py
@@ -1,0 +1,3 @@
+from .base import TimelineComponent
+from .pointlike import PointLikeTimelineComponent
+from .segmentlike import SegmentLikeTimelineComponent
